feat(api): add PrecomputedConfigParser interface and BasePrecomputedClient
Create the framework interfaces for precomputed client functionality:

- Add PrecomputedConfigParser<T> interface for JSON parsing abstraction
- Add PrecomputedParseException for parse error handling
- Create BasePrecomputedClient<T> abstract class with core logic
- Refactor EppoPrecomputedClient to extend BasePrecomputedClient<JsonNode>

This enables framework consumers to provide custom JSON implementations
while the batteries-included artifact provides Jackson defaults.
